experiments/toy_finetune/requirements.txt
torch>=1.13
transformers>=4.30
datasets
pyyaml
numpy
tqdm
